Update README.md
This is the first and only pirate album on the unique NFT 10,000 blockchain. Be quick guys because I have much bigger plans. The best is the right of all of us. Thank you for being my friend and a member of my family.
I am donating 20% of the sales of this album to help people